Use Puppet 4 as provisioner and in profiles, use official Icinga Puppet modules, rename scenarios

+++ modules/mongodb/lib/puppet/provider/mongodb_database/mongodb.rb
@@ -1,17 +1,16 @@
 require File.expand_path(File.join(File.dirname(__FILE__), '..', 'mongodb'))
-Puppet::Type.type(:mongodb_database).provide(:mongodb, :parent => Puppet::Provider::Mongodb) do
+Puppet::Type.type(:mongodb_database).provide(:mongodb, parent: Puppet::Provider::Mongodb) do
+  desc 'Manages MongoDB database.'
 
-  desc "Manages MongoDB database."
-
-  defaultfor :kernel => 'Linux'
+  defaultfor kernel: 'Linux'
 
   def self.instances
     require 'json'
     dbs = JSON.parse mongo_eval('printjson(db.getMongo().getDBs())')
 
-    dbs['databases'].collect do |db|
-      new(:name   => db['name'],
-          :ensure => :present)
+    dbs['databases'].map do |db|
+      new(name: db['name'],
+          ensure: :present)
     end
   end
 
@@ -19,22 +18,28 @@ Puppet::Type.type(:mongodb_database).provide(:mongodb, :parent => Puppet::Provid
   def self.prefetch(resources)
     dbs = instances
     resources.keys.each do |name|
-      if provider = dbs.find { |db| db.name == name }
-        resources[name].provider = provider
-      end
+      provider = dbs.find { |db| db.name == name }
+      resources[name].provider = provider if provider
     end
   end
 
   def create
-    mongo_eval('db.dummyData.insert({"created_by_puppet": 1})', @resource[:name])
+    if db_ismaster
+      mongo_eval('db.dummyData.insert({"created_by_puppet": 1})', @resource[:name])
+    else
+      Puppet.warning 'Database creation is available only from master host'
+    end
   end
 
   def destroy
-    mongo_eval('db.dropDatabase()', @resource[:name])
+    if db_ismaster
+      mongo_eval('db.dropDatabase()', @resource[:name])
+    else
+      Puppet.warning 'Database removal is available only from master host'
+    end
   end
 
   def exists?
-    !(@property_hash[:ensure] == :absent or @property_hash[:ensure].nil?)
+    !(@property_hash[:ensure] == :absent || @property_hash[:ensure].nil?)
   end
-
 end
